% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorocs.R
\name{sorocs}
\alias{sorocs}
\title{A Bayesian nonparametric Dirichlet process mixtures to Eestimate Receiver Operating Characteristic (ROC) Surface model}
\usage{
sorocs(
  nsim = 4,
  nburn = 2,
  Yvariable1,
  Yvariable2,
  gridY = seq(0, 5, by = 0.05),
  Xvariable1,
  Xvariable2
)
}
\arguments{
\item{nsim}{Number of simulations}

\item{nburn}{Burn in number}

\item{Yvariable1}{Dependent variable at setting 1}

\item{Yvariable2}{Dependent variable at setting 2}

\item{gridY}{a regular sequence spanning the range of Y variable}

\item{Xvariable1}{independent variable at setting 1}

\item{Xvariable2}{independent variable at setting 2}
}
\value{
A list of posterior estimates
}
\description{
A Bayesian nonparametric Dirichlet process mixtures to estimate the receiver operating characteristic (ROC) 
surfaces and the associated volume under the surface (VUS), a summary measure similar to the area under the 
curve measure for ROC curves. To model distributions flexibly, including their skewness and multi-modality
characteristics a Bayesian nonparametric Dirichlet process mixtures was used. Between-setting correlations is handled
by dependent Dirichlet process mixtures that have bivariate distributions with nonzero
correlations as their bases. To  accommodate ordering constraints, the stochastic ordering in the
context of mixture distributions was adopted.
}
\examples{
library(MASS)
library(MCMCpack)
library(mvtnorm)
data(sampledata)
try1 <- sorocs:::sorocs(Yvariable1 =asrm$logREscoremean2, Yvariable2=asrm$logREscoremean1, 
gridY=seq(0,5,by=0.5), Xvariable1=asrm$TN12/asrm$JN12 , Xvariable2 =asrm$TNN12/asrm$JNN12)
}
\keyword{Dirichlet}
\keyword{ROC}
\keyword{VUS}
\keyword{mixtures}
\keyword{process}
